- refactor + optimize coregame

+++ Assets/Scripts/Item Script/Booster.cs
@@ -13,10 +13,10 @@ public class Booster
 {
     public BoosterEffectType effectType;
     public Sprite boosterSpr;
-    public float value;
+    public int value;
     public float rate;
 
-    public Booster(BoosterEffectType effectType, float value, float rate)
+    public Booster(BoosterEffectType effectType, int value, float rate)
     {
         this.effectType = effectType;
         this.value = value;
